+++ Tests/Basic/Basic.cpp
@@ -1,0 +1,8 @@
+/*
+The Heady library is distributed under the MIT License (MIT)
+https://opensource.org/licenses/MIT
+See LICENSE.TXT or Heady.h for license details.
+Copyright (c) 2018 James Boer
+*/
+
+#include "Basic.h"