+++ Changelog.md
@@ -1,6 +1,12 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [0.2.0] - 2018-09-15
+
+- Added inline macro substitution to solve multiple definition errors
+- Changed API interface to use a single struct for parameters
+- Added basic self test that will catch multiple definition errors
+
 ## [0.1.0] - 2018-09-05
 
 - Uses a more robust header generation algorithm